Enable record generation in the debugger records.
After testing, recursive records such as:

arec = RECORD
  next:  POINTER TO arec;
END;

caused a problem since next has an undefined type and is skipped.  Since it's the last field in the record and it's skipped, the record doesn't have a trailing field which causes problems.

(Trying to access .next will cause a compiler error because it's undefined so I guess that isn't supported in Modula/2).

arecPtr = POINTER TO arec;
arec = RECORD
  next:  arecPtr;
END;

Undefined pointers are now treated as POINTER TO CHAR for debug record generation purposes.  This is consistent with how function pointers are treated.  (ORCA/C generates void * and function pointers as a POINTER TO INTEGER)